removed old configuration files, renamed new ones
