+++ tests/data/test_config.yml
@@ -1,19 +1,21 @@
-definitions:
-  creds_path: "../../credentials.json"
-  token_pickle: "token.pickle"
-  synapse_config: "../../.synapseConfig" ### Note: this key is required for people who use Synapse token authentication approach. 
-  service_acct_creds: "../../schematic_service_account_creds.json" ## Note: this key is required for google drive services 
+asset_store:
+  synapse:
+    config_basename: "file_name"
+    manifest_basename: 'file_name'
+    master_fileview_id: 'syn1'
+    manifest_folder: 'folder_name'
 
-synapse:
-  master_fileview: "syn23643253"
-  manifest_basename: "synapse_storage_manifest"
-  manifest_folder: 'manifests'
+manifest:
+  title: 'title'
+  data_type:
+    - 'data_type'
 
 model:
-  input:
-    location: 'example.model.jsonld'
-    file_type: 'local'
+  location: 'model.jsonld'
+  file_type: 'not_local'
 
-style:
-  google_manifest:
-    master_template_id: '1LYS5qE4nV9jzcYw5sXwCza25slDfRA1CIg3cs-hCdpU'
+google_sheets:
+  service_acct_creds_synapse_id: 'syn1'
+  service_acct_creds_basename: "creds.json"
+  master_template_id: 'id'
+  strict_validation: false
