--- tests/data/test_config2.yml
@@ -1,21 +1,0 @@
-asset_store:
-  synapse:
-    config_basename: "file_name"
-    manifest_basename: 'file_name'
-    master_fileview_id: 'syn1'
-    manifest_folder: 'folder_name'
-
-manifest:
-  title: 'title'
-  data_type:
-    - 'data_type'
-
-model:
-  location: 'model.jsonld'
-  file_type: 'not_local'
-
-google_sheets:
-  service_acct_creds_synapse_id: 'syn1'
-  service_acct_creds_basename: "creds.json"
-  master_template_id: 'id'
-  strict_validation: false
